requirements.txt
albumentations==1.1.0
livelossplot==0.5.4
matplotlib==3.4.3
notebook==6.4.4
numpy==1.21.*
opencv-python==4.5.3.*
pandas==1.3.3
Pillow==8.3.2
pytorch-lightning==1.4.9
scipy==1.7.1
torch==1.9.1
torchmetrics==0.5.1
torchvision==0.10.1
